+++ orderbook.cpp
@@ -4,7 +4,7 @@
 #include <mutex>
 #include "io.hpp"
 
-int Orderbook::length() {
+size_t Orderbook::length() {
   return book.size();
 }
 
@@ -25,22 +25,22 @@ void Orderbook::remove(int index) {
 }
 
 bool Orderbook::removeById(int id) {
-  for(int i = 0; i<(int)book.size(); i++){
+  for(size_t i = 0; i<(int)book.size(); i++){
     if (get<2>(book[i]) == id) {   
       std::lock_guard<std::mutex> lk(mut);
-	    book.erase(book.begin() + i);
+	    book.erase(book.begin() + (long)i);
       return true;
     }
   }
   return false;
 }
 
-void Orderbook::incrementExId(int index) {
+void Orderbook::incrementExId(size_t index) {
   std::lock_guard<std::mutex> lk(mut);
   get<3>(book[index]) += 1;
 }
 
-void Orderbook::decrementCount(int index, int numSubtracted) {
+void Orderbook::decrementCount(size_t index, int numSubtracted) {
   std::lock_guard<std::mutex> lk(mut);
   get<1>(book[index]) -= numSubtracted;
 }
@@ -68,26 +68,26 @@ switch (cmd) {
       // Loop through the sell book vector and find the lowest seller
       {
         std::lock_guard<std::mutex> lk(mut);
-        for(int i = length()-1; i>=0; i--) {
+        for(size_t i = length()-1; i>=0; i--) {
           if (get<0>(book[i]) <= sellPrice) {
             sellPrice = get<0>(book[i]);
-            bestIndex = i;
+            bestIndex = (int)i;
           }
         }
       }
       // If we found a seller...
       if (bestIndex != -1) {
-        incrementExId(bestIndex);
+        incrementExId((size_t)bestIndex);
         std::lock_guard<std::mutex> lk(mut);
         // If we want to buy more than we're selling, lower our count and remove the sell order
-        if (count >= get<1>(book[bestIndex])) {
-	        count -= get<1>(book[bestIndex]);
-          Output::OrderExecuted(get<2>(book[bestIndex]), activeId, get<3>(book[bestIndex]), get<0>(book[bestIndex]), get<1>(book[bestIndex]), getCurrentTimestamp());
+        if (count >= get<1>(book[(size_t)bestIndex])) {
+	        count -= get<1>(book[(size_t)bestIndex]);
+          Output::OrderExecuted((uint32_t)get<2>(book[(size_t)bestIndex]), (uint32_t)activeId, (uint32_t)get<3>(book[(size_t)bestIndex]), (uint32_t)get<0>(book[(size_t)bestIndex]), (uint32_t)get<1>(book[(size_t)bestIndex]), getCurrentTimestamp());
 	        remove(bestIndex);
           // Otherwise, set our count to 0 and lower the count of the sell order
         }  else {
-          decrementCount(bestIndex, count);
-          Output::OrderExecuted(get<2>(book[bestIndex]), activeId, get<3>(book[bestIndex]), get<0>(book[bestIndex]), count, getCurrentTimestamp());
+          decrementCount((size_t)bestIndex, count);
+          Output::OrderExecuted((uint32_t)get<2>(book[(size_t)bestIndex]), (uint32_t)activeId, (uint32_t)get<3>(book[(size_t)bestIndex]), (uint32_t)get<0>(book[(size_t)bestIndex]), (uint32_t)count, getCurrentTimestamp());
           count = 0;
         }
         // Return 0 if our order is sold, or how many we still need to buy
@@ -103,26 +103,26 @@ switch (cmd) {
       // Loop through the vector to find the highest seller
       {
         std::lock_guard<std::mutex> lk(mut);
-        for(int i = length()-1; i>=0; i--) {
+        for(size_t i = length()-1; i>=0; i--) {
           if (get<0>(book[i]) >= buyPrice) {
             buyPrice = get<0>(book[i]);
-            bestIndex = i;
+            bestIndex = (int)i;
           }
         }
       }
       // If we found a buyer...
       if (bestIndex != -1) {
-        incrementExId(bestIndex);
+        incrementExId((size_t)bestIndex);
         std::lock_guard<std::mutex> lk(mut);
         // If we are selling more than they are buying, remove the buyer and lower our sell count
-        if (count >= get<1>(book[bestIndex])) {
-          Output::OrderExecuted(get<2>(book[bestIndex]), activeId, get<3>(book[bestIndex]), get<0>(book[bestIndex]), get<1>(book[bestIndex]), getCurrentTimestamp());
-          count -= get<1>(book[bestIndex]);
+        if (count >= get<1>(book[(size_t)bestIndex])) {
+          Output::OrderExecuted((uint32_t)get<2>(book[(size_t)bestIndex]), (uint32_t)activeId, (uint32_t)get<3>(book[(size_t)bestIndex]), (uint32_t)get<0>(book[(size_t)bestIndex]), (uint32_t)get<1>(book[(size_t)bestIndex]), getCurrentTimestamp());
+          count -= get<1>(book[(size_t)bestIndex]);
 	        remove(bestIndex);
         // Otherwise, set our count to 0 and subtract our count from the buyers order
         }  else {
-          Output::OrderExecuted(get<2>(book[bestIndex]), activeId, get<3>(book[bestIndex]), get<0>(book[bestIndex]), count, getCurrentTimestamp());
-          decrementCount(bestIndex, count);
+          Output::OrderExecuted((uint32_t)get<2>(book[(size_t)bestIndex]), (uint32_t)activeId, (uint32_t)get<3>(book[(size_t)bestIndex]), (uint32_t)get<0>(book[(size_t)bestIndex]), (uint32_t)count, getCurrentTimestamp());
+          decrementCount((size_t)bestIndex, count);
 	        count = 0;
         }
       }
